move ShaderCategory to its own header
